Addresses nullable warnings in test projects
Enables nullable reference types in the test project and updates test classes to address nullable warnings.

This change ensures that the test suite is robust against null reference exceptions,
improving the overall quality and reliability of the codebase.

Also renames a test class for clarity.
