+++ src/Servus.Core.Tests/Servus.Core.Tests.csproj
@@ -6,6 +6,8 @@
     <IsPackable>false</IsPackable>
 
     <LangVersion>12</LangVersion>
+
+    <Nullable>enable</Nullable>
   </PropertyGroup>
 
   <ItemGroup>
